Move reward modifications to wrappers
This makes extending and combining modifiers easier. For instance for
future modifiers like using the sharp ratio as return signal it would be
much easier now to combine them with the relative or only final return
modifiers. Also it follows the concepts provided by OpenAI gym.
